gdb/mrk3: Alternative sorting of linetable for inline calls
Rework the sorting of the line table for the case where there are inline
function entries in the line table.  It is possible that after the
initial sorting (which sorts by pc then line number) that entries for
inline function calls could be in the middle of a set of entries for a
given address.

After the rewrite, inline function entries are sorted to the start of
the set of entries for a given address.

gdb/ChangeLog:

	* buildsym.c (maybe_record_line): Remove stray debug print out.
	(compare_line_numbers_for_inline): New function.
	(reorder_linetable_entries): Use qsort and
	compare_line_numbers_for_inline.