mrk3/bfd: Don't adjust symbol addresses from executables
If using a fully linked object as input to a link, then don't adjust the
symbol address when loading the input file, doing so will result in the
symbol address becoming incorrect.  The value of a symbol in a fully
linked ELF is already the absolute address of the symbol.

bfd/ChangeLog:

	* elflink.c (elf_link_input_bfd): Only adjust symbol addresses
	for symbols not loaded from executables.