mrk3/ld: Display additional warning about error locations being wrong
When the linker displays the file and line number information for an
error, if linker relaxation is turned on then the location given can be
wrong, this will often confuse users.

Sadly, fixing this issue could be tricky, and right now is lower
priority, so, as a temporary measure, add an additional warning that
will be displayed when the linker exits, if a warning or error has been
shown that includes location information, and if linker relaxation is
turned on.  This extra warning tells the user that the location
information in the other linker errors/warnings might not be correct,
and how they should disable linker relaxation if they would like more
accurate location information.

ld/ChangeLog:

	* ldmain.h (possibly_incorrect_error_message): Declare.
	* ldmain.c (possibly_incorrect_error_message): Define.
	(ld_cleanup): Possibly display additional warning message.
	* ldmisc.c (vfinfo): Set possibly_incorrect_error_message if
	appropriate.