gdb/remote: Don't use vKill if multi-process features are disabled
The gdb remote protocol documentation is clear that the vKill command
should not be used unless the multi-process feature is reported as
supported by the remote target.

Currently within gdb we check to see if the vKill packet is enabled or
not before using the vKill command, however, the only way to disable
vKill is from the gdb console, the result is that vKill will be sent to
targets that don't support it, and never claimed to support it.

After this commit I guard use of vKill with a check to see if the
multi-process feature is enabled or not.  I have removed the ability to
disable vkill specifically from the console, the user must now disable
the whole multi-process feature set as one.

I did consider leaving the separate vKill control switch in addition to
the multi-process control switch, but this seemed unnecessary, and I
worried that in the future another bug could be introduced where
PACKET_vKill was used to guard sending a vKill.

Removed:

  set remote kill-packet (auto|on|off)
  show remote kill-packet

Instead Use:

  set remote multiprocess-feature-packet (auto|on|off)
  show set remote multiprocess-feature-packet

gdb/ChangeLog:

	* remote.c (PACKET enum): Remove PACKET_vKill.
	(remote_kill): Use remote_multi_process_p.
	(remote_vKill): Use PACKET_multiprocess_feature.
	(_initialize_remote): Remove vKill configuration.
	* NEWS: Mention removed commands.