bfd: Use only sym value when patching localised relocs a second time.
When we are reprocessing relocations (i.e. linking something that has
been linked with --emit-relocs), only use the stored symbol value rather
than the section address.

This solves the issue where a localized symbol is not at the base
address of a section, and where the value includes address space bits.
In this case, we use the symbol value, as this already includes the
offset that would otherwise be discarded when using the section address.

bfd/ChangeLog:

	* elf64-mrk3.c (mrk3_elf_relocate_section): For localized symbols,
	use the symbol value and discard the section address.