MRK3: Update assertion checking if sym ends part way through an instruction
The assertion previously looked as follows:

if (sym_hash->root.u.def.value <= addr
    && (sym_hash->root.u.def.value + sym_hash->size > addr))
  {
    /* If this assert fires then we have a symbol that ends
       part way through an instruction.  Does that make
       sense?  */
    BFD_ASSERT (sym_hash->root.u.def.value + sym_hash->size
                >= addr + count);
    sym_hash->size -= count;
  }

This assertion checked to make sure that no symbol ended in the middle
of the set of bytes which were just deleted. It asserts that if a symbol
begins before the deleted bytes, that it doesn't also finish within the
deleted bytes.

Consider the following example where A, B and C are instructions and 1,
2 and 3 are symbols.

AAAA-BBBB-CCCC
       XX          <-- bytes to be deleted
1111
2222-2222-2222
3333-333

1) Does not trigger an assertion, it begins and ends before the set of
   deleted bytes.
2) Does not trigger an assertions, it begins before the deleted bytes,
   but also ends after them.
3) This triggers an assertion, as the symbol begins at or before the
   deleted bytes, and also finishes before the ends of those delete
   bytes.

The assumption of the assertion is that if you are relaxing an
instruction, when you relax and delete the extraneous bytes, no symbols
should finish in the middle of the deleted sequence. If that happens,
then it means that the symbol did not finish at an instruction boundary,
and could mean that the symbol is not the correct size.

This assertion was failing, when bytes were deleted when handling .align
directives. Consider this example, where A and B are *symbols* (not
instructions)

AAAAAAAABB

During relaxation, a number of instructions in A are shortened, leaving
space to be deleted.

AAAA----BB
    XXXX           <-- bytes to be deleted

The bytes are deleted, and B is moved to fill the space
AAAABB----
    XXXX

This circumstance triggers the same assertion, because B begins at the
same position as the sequence of deleted bytes but finishes before the
end of the sequence. This can happen whenever B is shorter than the
number of bytes which can be deleted before it.

The only change that this commit makes is to disable the assertion when
bytes are being deleted for the purpose of alignment.

bfd/ChangeLog.MRK3:

	* elf64-mrk3.c (mrk3_elf_relax_delete_bytes): Selectively disable an
	assertion which check if a symbol ended in the middle of a sequence
	of deleted bytes.