gdb: Fix lval for struct/struct-ptr expression evaluation
The lval-type for struct of struct-ptr values returned when
EVAL_AVOID_SIDE_EFFECTS is used is wrong.  GDB forces the lval-type to
be not_lval.  This commit sets the correct lval-type instead.

The changes in this commit are pulled from commits that either have
already been applied upstream, or have been posted to the mailing list,
but not yet applied.