gdb: Fix error checking on array indexing
This changes some Allinea code, but I think that the code now reflects
the comment that was originally in place.  The comment suggests that the
array index check should not be applied for single-dimension arrays, but
the code that was originally there turned off the error check for the
last dimension of all arrays, including multi-dimensional arrays.

After this commit, the error is disabled for just single-dimension
arrays, this fixes test in 'gdb.fortran/multi-dim.exp'.