gdb: Disassemble opcodes printing and layout changes
When printing the instructions for MRK3 disassembly, we previously
included the opcodes in the output.  This was a mistake as the opcodes
printing is handled by core GDB and including the opcodes with the
instructions just causes confusion.

As part of fixing this I have added code to improve the alignment of the
disassembly output (for CLI GDB) when also printing the opcodes (use
disassemble /r).  The alignment is based on the maximum instruction
width, which I now supply for MRK3 GDB.

The patch to fix the disassembly alignment is in core GDB, but is not
suitable for upstream GDB.  Targets such as x86-64 have a huge maximum
instruction size, which would push the disassembly output far to the
right, the average (or usual) instruction size on x86-64 is much less
than the maximum.  For MRK3 this problem does not exist.  There is a
patch on the mailing list that solves this disassembly formatting issue
in a much better way, but the patch is not ready for merging yet, so I
have added my quick MRK3 only solution for now.

gdb/ChangeLog:

	* mrk3-tdep.c (mrk3_fancy_print_insn): Extend header comment,
	don't print opcode bytes.
	(mrk3_gdbarch_init): Set maximum instruction width.
	* disasm.c (gdb_pretty_print_insn): Space opcodes and instructions
	based on maximum instruction width.