gdb: Avoid trailing whitespace in disassembly output
For some instructions we have supplementary information (for example
branches we display the actual destination address), however, for
instruction without this information we currently leave trailing
whitespace in the output.

This trailing whitespace is not really a problem, it's just annoying
when looking at raw MI trace output.  Plus its nice to avoid trailing
whitespace when possible.  Given that the fix is trivial, that's what
I've done in this commit.

gdb/ChangeLog:

	* mrk3-tdep.c (mrk3_fancy_print_insn): Avoid trailing whitespace
	when there's no supplementary instruction information.