ld/mrk3: Don't try to compute output section VMA too early
When the user makes use of attribute ((location ("ADDRESS))) on some
data in their source code we place the data into a section called
.mrk3.location.ADDRESS.  The linker then creates a new section called
.mrk3.location, the address of which is built from ADDRESS plus some
flag bits.

The source of these flag bits can be any one of the following:

  1. The user can supply explicit flag bits on the linker command line.

  2. The user can supply the name of some other output section from
  their linker script, the linker then reuses the flag bits from that
  output section to use as the flag bits on the .mrk3.location section
  it is creating.

  3. If non of the above have been used, then the linker will look for a
  .data section, and use the flag bits from that.

  4. If non of the above can be used the linker gives a warning, and
  uses 0 as the flag bits.

For options #2 and #3 the linker needs to extract the flag bits from the
address of an existing output section.  If the address is an integer
expression then we can simply convert the address to a VMA, and extract
the flag bits.

However, if the address is not a simple integer, but is instead a more
complex expression (and it doesn't have to be very complex, the linker
does not do any kind of constant folding during expression creation)
then the address can't be converted to a VMA and the linker will fail.

The solution is to make the address of the new .mrk3.location section be
an expression, this expression takes the expression that is the address
of the original data section, extracts the flags, and merges them with
the ADDRESS.

The linker will then resolve this expression later on in the linking
process in the normal way.

ld/ChangeLog:

	* emultempl/mrk3.em (handle_special_placement_sections): Don't
	attempt to convert output section origin expression to a VMA too
	early.