ld/bfd: Support relinking fully linked executables
When relinking fully linked ELF files the values held in symbols have
been converted to the absolute address for the symbol.  This is fine in
most cases as during a second link we don't expect to be merging any
sections together (and thus changing the sections address)....

.... except for debug sections, which we do join together.  To resolve
this issue, we now recalculate the offset of the symbol in the input
section and use this to compute the relocation value.

bfd/ChangeLog:

	* elf64-mrk3.c (mrk3_elf_relocate_section): Calculate symbol
	offset within input section.