ld: Warn when flags on input sections don't match flags on output section
In general the flags on input sections should match the flags on the
output section.  This commit gives a warning when this is not the case.

This is a warning rather than an error as the current shared library
solution does rely on changing the flags during the second link phase.
It would be nice, once the shared library solution is understood a
little better, if the warning could be silenced in this case, however,
for now the warning can be ignored for shared libraries.

The hope is that this warning will be useful as a diagnostic aid in the
future if we have any linker issues.

ld/ChangeLog:

	* emultempl/mrk3.em (gld${EMULATION_NAME}_finish): Warn when flags
	on input sections don't match the flags on output sections.