ld: Override default linker emulation at configure time
This commit adds a new configure option for ld that allows the default
linker emulation to be selected at configure time.  The new option is:

    --with-default-linker-emulation=NAME

where NAME is the name of a linker emulation that would otherwise be
valid for this target (as you'd pass to ld -mNAME).  The validity of
NAME is checked at configure time.

The effect of this configure option is to make NAME the new default
linker emulation.  The old default is still available in the list of
alternative emulations and can be selected with the -m option to ld.

If this new option is not used at configure time then the linker should
be unchanged.

ld/ChangeLog:

	* configure.ac: Add option --with-default-linker-emulation.
	* configure: Regenerate
	* NEWS: Add mention of new option.