ld: Make SORT_BY_* and EXCLUDE_FILE separate rules in the grammar
Split the rules that handle the SORT_BY_* and EXCLUDE_FILE for
sections.  This split means that EXCLUDE_FILE is now supported in more
places, making its handling much more consistent.

For example, this was valid:

  *(SORT_BY_NAME ( EXCLUDE_FILE ( a.o ) .text.* ))

but this is not:

  *(SORT_BY_NAME ( SORT_BY_NAME ( EXCLUDE_FILE ( a.o ) .text.* ) ))

despite this being legal:

  *(SORT_BY_NAME ( SORT_BY_NAME ( .text.* ) ))

Or this is not legal:

  *(SORT_BY_ALIGNMENT ( SORT_BY_NAME ( EXCLUDE_FILE (a.o) .text.* )))

despite this being legal:

  *(SORT_BY_ALIGNMENT ( SORT_BY_NAME ( .text.* )))

After this commit all of these inconsistencies are resolved, anywhere
that we can give a section wildcard we can now prefix that with
EXCLUDE_FILE.

ld/ChangeLog:

	* ldgram.y (section_name_spec): Simplifiy and use
	wildcard_maybe_exclude.
	* testsuite/ld-scripts/exclude-file-5.d: New file.
	* testsuite/ld-scripts/exclude-file-5.map: New file.
	* testsuite/ld-scripts/exclude-file-5.t: New file.
	* testsuite/ld-scripts/exclude-file-6.d: New file.
	* testsuite/ld-scripts/exclude-file-6.map: New file.
	* testsuite/ld-scripts/exclude-file-6.t: New file.
	* NEWS: Mention change.