gdb/arc: Use stop_pc from sigcontext to unwind the pc
The table arc_linux_sc_reg_offset is actually indexed by gdb register
number, and used to initialise the unwind location of each register
within a signal frame.

The index in this table that is used to initialise the unwind location
of the pc (which is marked as stop_pc in the table) should reference the
location of the stop_pc field within the sigcontext structure that will
be on the stack for the signal frame.

Currently this field is filled in as REGISTER_NOT_PRESENT, but by
providing a correct offset it should be possible to unwind the pc within
a signal frame, and thus backtrace beyond a signal.

gdb/ChangeLog:

	* arc-linux-tdep.c (arc_linux_sc_reg_offset): Fill in the pc field
	to refer to the stop_pc.