gdb/arc: Correct offset of sigcontext on stack frame
The arc_linux_sigcontext_addr function should return the address of the
sigcontext in the stack for a given frame (assuming the frame is a
signal frame).  We currently assume that the sigcontext is the first
thing on the stack, however, this is not true.  Signal frames are
actually passed a ucontext_t structure, a sigcontext is buried within
the structure, and is _not_ at offset 0.

I don't currently have a good solution for "computing" the offset of the
sigcontext on the stack frame, so for now, I've just hard coded the
offset into GDB.  There are obvious problems with this, if the offset
changes things will go wrong.  Worse, if different customers require
different offsets then we'll be in real trouble.

For now, just hard coding an offset that is correct _now_ seems better
than sticking with the 0 offset that we know is wrong.  With this change
we can now find the sigcontext correctly for signal frames.

gdb/ChangeLog:

	* arc-linux-tdep.c (arc_linux_sigcontext_addr): Fix offset of
	sigcontext.