gdb/arc: Improvement to prologue scanner
If we spot a st.a instruction, and the memory offset is -4, but we're
not storing blink or fp register, then we should still perform the check
to see if it's a callee saved register, and then adjust the stack
offset.

We previously cut the prologue scan short in random cases when we hit a
st.a to the stack with offset -4, but were not storing blink or fp
register.

gdb/ChangeLog:

	* arc-tdep.c (arc_is_in_prologue): Check for st.a to stack for
	case where offset is -4, but we're not pushing blink or fp.