gdb/arc: Don't stop prologue scan for call instructions
In some cases the prologue can contain call instructions to write out
registers, when this is the case we should not terminate the prologue
scan at that call instruction.

This commit is a first cut at achieving this aim, though this commit
might be a little too aggressive.  Currently with this commit we only
stop at some non-call branch instructions, we should probably change
this to detect the specific call instructions that can be found in a
prologue, and accept them.

TODO: Rewrite this commit to accept less instructions.

gdb/ChangeLog:

	* arc-tdep.c (arc_is_in_prologue): Only stop at some branch
	instructions.