ld: Use full names for SORT_* specifiers in map files
When creating map files all SORT_* specifiers (SORT_BY_NAME,
SORT_BY_ALIGNMENT, etc) are all printed as just 'SORT'.  The result is
that in this area the map file does not present a full record of what
was going on.

After this commit the full name of the sort specifier is now used in
the map file.  I checked through the testsuite, and couldn't find any
tests covering this area.

As 'SORT' is an alias for 'SORT_BY_NAME' this does mean that some users
might be surprised by this change.  However, I felt that changing SORT
to SORT_BY_NAME was less confusing than changing SORT_BY_NAME to SORT.

In the case of filename lists where only SORT_BY_NAME (or its alias
SORT) are valid it could leaving things as they are at the
moment (always displaying 'SORT') would have been an option, however, I
changed to 'SORT_BY_NAME' because I thought consistency with the section
lists was a good thing, and we might as well make all the change in this
area in one go.

ld/ChangeLog:

	* ldlang.c (print_wild_statement): Use full names for SORT
	specifiers.