WIP: bfd: Tighten section to segment mapping logic
The ELF_SECTION_IN_SEGMENT logic would allow in some cases a NOBITS
section to be mapped into two disjoint segments.

In the examples I'm seeing right now a NOBITS section will still have a
file offset within the ELF, but will have filesize of zero.  The code
tries to make sure that the file offset of the segment is inline with
the file offset of a section.

Consider this case:

ELF file: |____________________________________________|
                 |                       |
                 '---- NOBITS section    |
                       file offset       '--- Segment file
                                              offset

Currently the NOBITS section can be mapped into the segment so long as
the VMAs are close.  However, this section was probably not in the
segment when the ELF was created, and putting the NOTBITS section into
the segment can cause memory issues (out of memory) when trying to
objcopy the file.

I'm currently trying to come up with a good reproducer for this issue.