ld: Improve file-spec handling in the linker
The linker currently makes many calls to the walk_wild function which
has the task of searching through all lang_input_statement_type
instances and finding those whose filename matches the pattern within
a lang_wild_statement_type.

The problem is that this is done many times within a link, and can be
very expensive, however, the results should never change once all of
the input files are open and all of the linker created sections have
been created.

This patch changes the linker to cache the results of matching the
filenames against the lang_wild_statement_type patterns, then on each
subsequent iteration the results can be pulled from the cache, which
gives a reasonable performance boost to the linker.

ld/ChangeLog:

	* ldlang.c (walk_wild_add_match): New function.
	(walk_wild_init_matches): New function.
	(walk_wild_populate_matches): New function.
	(walk_wild): Rewrite to populate cache, and then walk matches out
	of the cache.
	(open_input_bfds): Update location of filename field.
	(print_wild_statement): Likewise.
	(lang_add_wild): Update location of filename field, and initialise
	new matches field.
	* ldlang.h (struct lang_input_statement_list): New structure.
	(struct lang_filename_spec): New structure.
	(struct lang_wild_statement_struct): Remove 'filename' field, and
	add 'filename_spec' field.