bfd: Improve value fetching from bfd contents
There are generic routines for fetching values of various sizes out of
the bfd image.  These routines fetch the value byte by byte and
assemble the final value using shift and bitwise `or' in order to have
generic code handle both big and little endian hosts.

The problem is that this code is less optimal when, for example,
fetching a little endian value on a little endian host, where we could
just copy the value directly out of memory using memcpy.

This commit introduces this optimisation, which makes a noticable
perfomance improvement, around 3%+ on moderately sized links.

bfd/ChangeLog:

	* libbfd.c (bfd_getl16): Optimise the case of running on a little
	endian host.
	(bfd_getl_signed_16): Likewise.
	(bfd_getl32): Likewise.
	(bfd_getl_signed_32): Likewise.
	(bfd_getl64): Likewise.
	(bfd_getl_signed_64): Likewise.