bfd: Improve value fetching from bfd contents
There are generic routines for fetching values of various sizes out of
the bfd image.  These routines fetch the value byte by byte and
assemble the final value using shift and bitwise `or' in order to have
generic code handle both big and little endian hosts.

Though this approach is fine, it is slower than simply performing a
direct copy using memcpy.

As all of this code is hidden away in libbfd.c, this commit applies
some optimisation; if the host and target are the same endianness then
we now use memcpy instead of shift and `or'.

With this optimisation in place I am seeing around a 3% speed increase
in moderately sized links.

bfd/ChangeLog:

	* libbfd.c (bfd_getl16): Optimise the case of running on a little
	endian host.
	(bfd_getl_signed_16): Likewise.
	(bfd_getl32): Likewise.
	(bfd_getl_signed_32): Likewise.
	(bfd_getl64): Likewise.
	(bfd_getl_signed_64): Likewise.