bfd: Don't repeatedly lookup arch information
Looking up the arch information involves a linear search of a list,
and can be done millions of times in a link of even moderate size.

The list being searched is pretty small, but even so, the
bfd_octets_per_byte function can end up accounting for 3% or more of a
moderately sized link.

Given that the information being looked up is already cached on the
bfd object it is trivial to optimise this function.

After optimisation this function drops to accound for less than 0.5%
of the link time.

bfd/ChangeLog:

	* archures.c (bfd_octets_per_byte): Use bits_per_byte on the
	cached arch_info instead of looking up the information each time.