gdb: Clear per-inferior data when an inferior exits
The per-inferior registry data is currently associated with each
inferior object, not with a program execution within that inferior
object.  The result of this is that given this sequence of commands:

  file prog1.exe
  start
  kill
  file prog2.exe
  start

there will only be one inferior object created, the default inferior,
that is created when GDB starts up.  Any per-inferior data that is
associated with this inferior will exist until GDB exits, including
through the execution of prog1.exe and prog2.exe.

Clearly this hasn't caused any problems so far, however, it does mean
that per-inferior data doesn't lend itself to holding information that
might change based on the program being executed, or the target the
inferior is running on, or at least, if we did want to hold data that
varies per program or target then we need to validate it on each use.

An alternative strategy would be to discard all per-inferior data as
soon as the inferior exists.  This would cause GDB to recreate the data
once a new program has started, possible on a new target.  Which is what
this commit does.

Tested on x86-64 GNU/Linux with no regressions.

gdb/ChangeLog:

	* inferior.c (exit_inferior_1): Clear per-inferior data when an
	inferior exits.