gdb/riscv: fix issues running 32-bit float on 64-bit float target
NOTE: This commit is not good, and should not be pushed upstream.
Running a binary compiled for 32-bit float on a machine with 64-bit
float, is, apparently, invalid.  I am preserving this commit for now
as I might want to pick through this for any other useful changes.

When running a binary compiled for rv32imf on an rv32imdf machine then
GDB makes some incorrect choices about how arguments are passed when
making an inferior call.

The reason is that we make abi related choices based on the available
hardware features, rather than the abi declared in the bfd file being
tested.

This commit splits the existing feature check methods into two sets,
one that checks available hardware features, and one that checks abi
features.

With this change in place the gdb.base/callfuncs.exp test does much
better when compiled for rv32imf and run on rv32imdf.

gdb/ChangeLog:

	* riscv-tdep.c (riscv_has_feature): Renamed to...
	(riscv_isa_has_feature): ...this.  Update comments.
	(riscv_abi_has_feature): New function.
	(show_use_compressed_breakpoints): Update for
	riscv_isa_has_feature name change.
	(riscv_isa_xlen): Now calls riscv_abi_xlen, old implementation
	moved to...
	(riscv_abi_xlen): ...this new function.
	(riscv_abi_flen): New function.
	(riscv_isa_flen): Update to call riscv_isa_has_feature.
	(riscv_has_fp_abi): Rename to...
	(riscv_abi_any_float_p): ...this.
	(riscv_breakpoint_kind_from_pc): Update to call
	riscv_isa_has_feature.
	(struct riscv_call_info::riscv_call_info): Get xlen and flen from
	abi details, not isa details.
	(struct riscv_call_info::xlen): Update comment.
	(riscv_push_dummy_call): Update to call riscv_abi_any_float_p.
	(riscv_gdbarch_init): Update due to changes in gdbarch_tdep
	structure.
	* riscv-tdep.h (struct gdbarch_tdep): Update comments, and move
	features bit-field into abi sub-struct.