gdb: enable [] to find children using pretty-printers
This commit allows GDB to make use of an available pretty-printer to
provide [] indexing into a type.

Consider the following example before any pretty-printer is loaded:

  (gdb) p obj_int
  $1 = {
    m_name = 0x4052a0 "first int array",
    m_nitems = 4,
    m_items = 0x4052c0
  }
  (gdb) p obj_int[3]
  Structure has no component named operator[].

Now, after loading a pretty-printer:

  (gdb) p obj_int
  $2 = int array "first int array" with 4 elements = {3, 6, 4, 2}
  (gdb) p obj_int[3]
  $3 = 2

This []-indexing will only occur for objects whose pretty-printers
have a display_hint of 'array' or 'map'.  And the pretty-printer based
indexing is only used if compiled in methods (for C++) or xmethods are
not available.

The benefit I see in this commit is that this feels like an easy win
for the user; they write the pretty printer and they get this extra
functionality for free.

The downside is the extra complexity within GDB.  I'm still not
completely sure if the benefit is worth the cost, and I'd love to hear
people's thoughts on this.

gdb/ChangeLog:

	* NEWS: Mention use of pretty-printer for [] indexing.
	* eval.c (evaluate_subexp_standard): Use pretty printers to find
	children for BINOP_SUBSCRIPT.
	* extension-priv.h (struct extension_language_ops)
	<val_pretty_printer_find_child>: New member variable.
	* extension.c (ext_lang_pretty_printer_find_child): New function.
	* extension.h (ext_lang_pretty_printer_find_child): Declare new
	function.
	* guile/guile.c (guile_extension_ops): Add new field, initialised
	to NULL.
	* python/py-prettyprint.c (gdbpy_val_pretty_printer_find_child):
	New function.
	* python/python-internal.h (gdbpy_val_pretty_printer_find_child):
	Declare.
	* python/python.c (python_extension_ops): Add new field, initialised
	to gdbpy_val_pretty_printer_find_child.

gdb/doc/ChangeLog:

	* python.texi (Pretty Printing API): Mention the use of
	pretty-printers for [] indexing.

gdb/testsuite/ChangeLog:

	* gdb.python/py-prettyprint-subscript.c: New file.
	* gdb.python/py-prettyprint-subscript.cc: New file.
	* gdb.python/py-prettyprint-subscript.exp: New file.
	* gdb.python/py-prettyprint-subscript.py: New file.