Additional tests for scheduler-locking related bugs
This test is derived from the test mentioned in this mailing list
post:

   https://sourceware.org/ml/gdb/2017-10/msg00032.html

The test starts a set of threads, turns on scheduler-locking, and then
makes inferior calls in each thread.  A second test starts the same
set of threads, and continues each thread to a second breakpoint.

gdb/testsuite/ChangeLog:

	* gdb.threads/schedlock-02.c: New file.
	* gdb.threads/schedlock-02.exp: New file.
	* lib/gdb.exp (get_current_thread_id): New function.