gdb: Seperate error message for entry value of an unknown symbol
This commit changes the error message that is produced when the user
asks for the entry value of an unknown symbol, hopefully making things
clearer for the user.

Currently, if asking for the entry value of an unknown symbol the user
gets this error message:

  @entry can be used only for function parameters, not for "<symbol>"

Which seems to emphasise <symbol> not being a function parameter.
After this patch the error message changes to:

  No symbol "<symbol>" in current context.

Which focuses on the fact that the symbol is unknown.

gdb/ChangeLog:

	* c-exp.y (variable: name_not_typename ENTRY): Give a separate
	error message for symbol not found case.  Rely on the
	SYMBOL_IS_ARGUMENT check to determine if a symbol is an argument.

gdb/testsuite/ChangeLog:

	* gdb.arch/amd64-entry-value.exp: Update error message for the
	unknown symbol case.