gdb: After inferior call, finish all threads in process
Currently, if non-stop mode is off and scheduler-locking is on, then
when we perform the inferior call we only call finish_thread_state on
the thread in which the call was made.

The problem with this is if the inferior call caused a new thread to
be created, the thread _will_ have been stopped when the inferior call
completed, but it's frontend state will be left as running (as
finish_thread_state was never called for it).

This change causes us to call finish_thread_state for all threads in
the process, which is what happens when scheduler-locking is off.
After this commit new threads, created in the inferior call, are
correctly shown as stopped once the inferior call has completed.

I did spend some time considering whether, when scheduler-locking is
on, whether it was even correct for the new thread to be allowed to
run at all (after it is created in the inferior call).  In the end I
decided that it probably was, if we imagine calling an inferior
function that creates a new thread, does some work, and then joins the
new thread again, we'd probably expect that to work, even with
scheduler-locking on.  As a result I've only conserned myself with the
state of new threads after the inferior call finishes.

gdb/ChangeLog:

	* infrun.c (run_inferior_call): Finish all threads in inferior
	process in order to catch possible new threads.

gdb/testsuite/ChangeLog:

	* gdb.threads/infcall-create-thread.c: New file.
	* gdb.threads/infcall-create-thread.exp: New file.